+++ tests/Endpoint/AccountEndpointTest.php
@@ -26,6 +26,7 @@ class AccountEndpointTest extends EndpointTestCase
         self::assertDefaultEndpointPath(self::endpoint()->signatureScenarios(), '/api/account/signature-scenarios');
         self::assertDefaultEndpointPath(self::endpoint()->identifyScenarios(), '/api/account/identify-scenarios');
         self::assertDefaultEndpointPath(self::endpoint()->groups(), '/api/account/groups');
+        self::assertDefaultEndpointPath(self::endpoint()->contacts(), '/api/account/contacts');
     }
 
     public function testGet(): void