+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes will be documented in this file.
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## [Unreleased]
+- Add `AccountEndpoint.contacts` CRUD endpoints
 - Add `MyEnvelopeRecipient.signingOrder` property
 - Add `GroupsEndpoint` endpoint
 - Add `GroupUsersEndpoint` endpoint